scripts updated.

+++ scripts/post_build.sh
@@ -2,5 +2,7 @@
 # post-build; written 2020 by Martin Zeitler
 
 shopt -s dotglob
-rm -r /root/.gradle/caches/*
 rm -r /workspace/*
+if [ -d "/root/.gradle/caches" ]; then
+    rm -r /root/.gradle/caches/*
+fi
